angle reading placed in it's own intterupt
the angle gets updated with it's own interrupt routine
- oversampling the angle with 2 time pid frequency
- the angle is updated even if the controller interrupt is switched off
-> makes (for example) the autotune function easier

+++ Mechaduino/Utils.h
@@ -27,7 +27,7 @@ void antiCoggingCal();
 
 void PID_autotune();
 
-extern int digitalSmooth(int rawIn, int *sensSmoothArray);
+extern float digitalSmooth(int rawIn, int *sensSmoothArray);
 
 bool canceled();
